Remove NeDB support
